example/qt-demo/pyproject.toml
[project]
name = "qt-demo"
version = "0.1.0"
description = "Example Qt (PySide6) desktop demo for uvpacker."
requires-python = "==3.12.*"
dependencies = [
    "PySide6-Essentials",
]

[project.scripts]
qt-demo = "qt_demo.main:main"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

